+++ docs/changelog.rst
@@ -67,6 +67,9 @@ Bug fixes:
 
 For plugin developers:
 
+- Metadata plugins can now implement a ``before_album_info_emitted`` method to
+  modify ``AlbumInfo`` objects before they are emitted as part of the
+  ``albuminfo_received`` event.
 - A new plugin event, ``album_matched``, is sent when an album that is being
   imported has been matched to its metadata and the corresponding distance has
   been calculated.